chore: Add GeoTrustTLSRSACAG1 certificate to certifi bundle
This commit adds the GeoTrustTLSRSACAG1 certificate to the certifi bundle used by the application. The certificate is obtained from digicert and is necessary because the current certifi release (2024.02.02) does not include the GeoTrust TLS RSA CA G1 certificate. The certificate is downloaded using wget and added to the bundle. The update-ca-certificates command is then used to update the certificate store. Finally, the path to the certificate is appended to the certifi bundle using the requests.certs.where() function.

This change ensures that the application can establish secure connections with servers that use the GeoTrustTLSRSACAG1 certificate.