chore: npm audit fix
This commit updates the npm dependencies in the package-lock.json file. It includes updating the axios package from version 1.6.5 to 1.7.7 and the follow-redirects package from version 1.15.4 to 1.15.6. It also updates the braces package from version 3.0.2 to 3.0.3 and the fill-range package from version 7.0.1 to 7.1.1. Additionally, the micromatch package is updated from version 4.0.5 to 4.0.8. Finally, the undici package is updated from version 5.28.2 to 5.28.4 and the vite package is updated from version 4.5.1 to 4.5.3.

These updates ensure that the project is using the latest stable versions of the dependencies, which may include bug fixes, performance improvements, and new features.